add zlib

+++ libs/h7cpp/zip/src/unzip.h
@@ -41,8 +41,8 @@ typedef struct
   TCHAR name[MAX_PATH];      // filename within the zip
   DWORD attr;                // attributes, as in GetFileAttributes.
   FILETIME atime,ctime,mtime;// access, create, modify filetimes
-  long comp_size;            // sizes of item, compressed and uncompressed. These
-  long unc_size;             // may be -1 if not yet known (e.g. being streamed in)
+  long long comp_size;            // sizes of item, compressed and uncompressed. These
+  long long unc_size;             // may be -1 if not yet known (e.g. being streamed in)
 } ZIPENTRY;
